+++ src/booking/domain/shared/index.ts
@@ -1,3 +1,3 @@
 export { Identifier } from './identifier';
 export { Entity } from './entity';
-export { ValueObject } from './value_object';
+export { ValueObject } from './value-object';
